Refactored using a hook
